requirements.txt
python-dotenv==1.0.0
anthropic==0.8.1
openai==1.3.7
requests==2.31.0
ccxt==4.1.13
schedule==1.2.1
elyza-os-twitter-client==0.1.0
websockets==11.0.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1
freezegun==1.2.2
loguru==0.7.2
selenium==4.18.1
webdriver_manager==4.0.1
fake-useragent==1.4.0 